Move `uri.GoogleCloudHealthcareURL` to new module `ext.gcp.uri`.
