+++ setup.py
@@ -42,13 +42,13 @@ setuptools.setup(
     package_dir={'': 'src'},
     extras_require={
         'gcp': [
+            'dataclasses>=0.8; python_version=="3.6"',
             'google-auth>=1.6',
             'google-oauth>=1.0',
         ],
     },
     python_requires='>=3.6',
     install_requires=[
-        'dataclasses; python_version=="3.6"',
         'pydicom>=2.0',
         'requests>=2.18',
         'retrying>=1.3.3',
