+++ src/dicomweb_client/ext/__init__.py
@@ -1,0 +1,1 @@
+"""Vendor-specific extensions of the `dicomweb_client` package."""
