+++ src/dicomweb_client/ext/gcp/__init__.py
@@ -1,0 +1,10 @@
+"""Google Cloud Platform (GCP) compatible extensions of `dicomweb_client`.
+
+Modules under this package may require additional dependencies. Instructions for
+installation are available in the Installation Guide here:
+https://dicomweb-client.readthedocs.io/en/latest/installation.html#installation-guide
+
+For further details about GCP, visit: https://cloud.google.com
+"""
+
+from dicomweb_client.ext.gcp.uri import GoogleCloudHealthcareURL
